% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{env}
\alias{env}
\title{bash alias for \code{\link[utils]{sessionInfo}}.}
\usage{
env(package = NULL)
}
\arguments{
\item{package}{a character vector naming installed packages, or \code{NULL}
   (the default) meaning all attached packages.}
}
\description{
bash alias for \code{\link[utils]{sessionInfo}}.
}

