% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.r
\docType{data}
\name{error_pkgs}
\alias{error_pkgs}
\title{List of packages that were not installed for some reason.}
\format{An object of class \code{NULL} of length 0.}
\usage{
error_pkgs
}
\description{
List of packages that were not installed for some reason.
}
\keyword{internal}

