% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.r
\docType{data}
\name{installed_pkgs}
\alias{installed_pkgs}
\title{List of packages already installed.}
\format{An object of class \code{NULL} of length 0.}
\usage{
installed_pkgs
}
\description{
List of packages already installed.
}
\keyword{internal}

