% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.r
\name{installer}
\alias{installer}
\title{Helper function that invokes the proper install function}
\usage{
installer(p)
}
\arguments{
\item{p}{character vector representing package name passed to \code{\link{install.packages}}.
If a list, the first element is the suffix to append to \code{devtools::install_}.
The second element is the path to the path package. Refer to \code{\link[devtools]{install}} for more info.}
}
\value{
returns a named list if successful.  Otherwise FALSE.
}
\description{
Helper function that invokes the proper install function
}
\section{Limitations}{

There are many params to the install functions that are not supported at this time.
}
\keyword{internal}

