% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.r
\docType{data}
\name{my_packages}
\alias{my_packages}
\title{Default set of packages.}
\format{An object of class \code{list} of length 21.}
\usage{
my_packages
}
\description{
To see a tidy list of them, refer to Examples:
}
\examples{
dgutils::fl(my_packages)

}
\keyword{datasets}

