% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_packages.r
\docType{data}
\name{new_pkgs}
\alias{new_pkgs}
\title{List of packages this function installed.}
\format{An object of class \code{NULL} of length 0.}
\usage{
new_pkgs
}
\description{
List of packages this function installed.
}
\keyword{internal}

