% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_package.R
\name{tailor_package}
\alias{tailor_package}
\title{Complete package setup after RStudio creates a new project.}
\usage{
tailor_package()
}
\description{
Complete package setup after RStudio creates a new project.
}
