% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{cd}
\alias{cd}
\title{bash alias for \code{\link[base]{setwd}}.}
\usage{
cd(dir)
}
\arguments{
\item{dir}{A character string: \link{tilde expansion} will be done.}
}
\description{
bash alias for \code{\link[base]{setwd}}.
}

