% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl.R, R/flp.R
\name{fl}
\alias{fl}
\alias{flp}
\title{An indented, wrapped list styled after a definition.}
\usage{
fl(x, label = "", indent_string = NA)

flp(x, label = "", indent_string = NA)
}
\arguments{
\item{x}{the R object to display. \code{\link{toString}} must be able to evaluate it.}

\item{label}{a string to describe the list.  The remaining lines are
indented by the number of characters from catcatenating \code{indent_string} and
`label`.}

\item{indent_string}{the string to prefix the first line.  Defaults to
\code{tab.width} option number of spaces.  If you wanted your
list indented 4 spaces, pass a string of 4 spaces to this parameter.}
}
\value{
a formatted string
}
\description{
There are many formatting or pretty printing functions available,
some of them quite elaborate.  Mine uses only functions in \code{base}.  It seems that many
programmers don't realize the power of \code{strwrap} and the \code{sep} parameter for \code{\link{cat}}.

An indented, wrapped list styled after a definition.
}
\section{Functions}{
\itemize{
\item \code{flp}: formatted string variant
}}
\section{R Trivia}{

Turns out the \code{tab.width} option is not set by default in base R as
I discovered when I tried to build the package.  Now explicitly test to see if the
option is set.
}
\examples{
fl(search(), label = "search():  ")

s <- flp(search(), label = "search():  ")

}

