% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-libraries.r
\name{installer}
\alias{installer}
\title{installer helper function that does the install}
\usage{
installer(p)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
installer helper function that does the install
}

