% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\title{Note a comment in the history file for future reference.}
\usage{
note(text = "Previous command worked")
}
\arguments{
\item{text}{The default is \code{"Previous command worked"}.}
}
\value{
The character string returned by \code{\link[utils]{timestamp}}.
}
\description{
Sometimes interactive exploration of running different R statements is a
great way to understand what is happening with your code or data.
\code{note} is a quick way to annote your results in the history file so you
can record what worked or did not work or a question that needs further
research.
}
\examples{
note()
note("Calculation matched reference data!")

}

