% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-libraries.r
\name{reinstall_packages}
\alias{reinstall_packages}
\title{reinstall_packages installs your standard set of packages}
\usage{
reinstall_packages(pkgs = NA)
}
\arguments{
\item{pkgs}{}
}
\description{
reinstall_packages installs your standard set of packages
}

