% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-ggplot.R
\docType{data}
\name{theme_dg}
\alias{theme_dg}
\title{Modifications to \code{theme_bw} to add a bit of color.}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
theme_dg
}
\description{
Modifications to \code{theme_bw} to add a bit of color.
}
\seealso{
\code{\link[ggplot2]{ggtheme}}
}
\keyword{datasets}
