% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{data}
\name{theme_dg}
\alias{theme_dg}
\title{theme_dg Modified theme_bw for a simple, colored version of the theme.}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 44.}
\usage{
theme_dg
}
\description{
theme_dg Modified theme_bw for a simple, colored version of the theme.
}
\keyword{datasets}

